+++ jaguar/lib/http/mux/route.dart
@@ -76,7 +76,7 @@ class Route {
   factory Route(String path, RouteHandler handler,
           {List<String> methods = const <String>['*'],
           Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor,
@@ -100,7 +100,7 @@ class Route {
   /// Constructs a [Route] for GET method requests
   factory Route.get(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor,
@@ -124,7 +124,7 @@ class Route {
   /// Constructs a [Route] for POST method requests
   factory Route.post(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor,
@@ -148,7 +148,7 @@ class Route {
   /// Constructs a [Route] for PUT method requests
   factory Route.put(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor,
@@ -196,7 +196,7 @@ class Route {
   /// Constructs a [Route] for PATCH method requests
   factory Route.patch(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor,
@@ -220,7 +220,7 @@ class Route {
   /// Constructs a [Route] for OPTIONS method requests
   factory Route.options(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor,
@@ -244,7 +244,7 @@ class Route {
   /// Constructs a [Route] for HTML requests
   factory Route.html(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType = MimeTypes.html,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor,
@@ -274,7 +274,7 @@ class Route {
             'DELETE'
           ],
           Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor = jsonResponseProcessor,
@@ -298,7 +298,7 @@ class Route {
   /// Constructs a [Route] for JSON requests with GET method
   factory Route.getJson(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor = jsonResponseProcessor,
@@ -322,7 +322,7 @@ class Route {
   /// Constructs a [Route] for JSON requests with POST method
   factory Route.postJson(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor = jsonResponseProcessor,
@@ -346,7 +346,7 @@ class Route {
   /// Constructs a [Route] for JSON requests with PUT method
   factory Route.putJson(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor = jsonResponseProcessor,
@@ -370,7 +370,7 @@ class Route {
   /// Constructs a [Route] for JSON requests with DELETE method
   factory Route.deleteJson(String path, RouteHandler handler,
           {Map<String, String>? pathRegEx,
-          int statusCode = 200,
+          int? statusCode,
           String? mimeType,
           String? charset = kDefaultCharset,
           ResponseProcessor? responseProcessor = jsonResponseProcessor,